api/cellar/yeast,cellar: Register yeast items
Yeast can be registered, but has no biome mapping, this along with a small
change to the Yeast Jar's produce logic allows the multiplication of hard to
get yeast types (Origin and Bayanus).

I may consider wrapping the Yeast entries in an actual class to keep things organized, but lists and maps work fine for now.